Basic engine refactoring. Moved game loop and base game logic into its own overridable class.
